+++ quill_native_bridge_macos/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.0.1-dev.4
+
+- Support Swift Package Manager.
+
 ## 0.0.1-dev.3
 
 - Require `quill_native_bridge_platform_interface` minimum version `0.0.1-dev.4`.